Update SiFive U74 raw event DT bindings
Previous patch from Nikita was for older version of OpenSBI, this
updates the bindings for U74 raw events to latest OpenSBI, tested
on v1.0+ but the related code changes happened months ago.

This update is from Nikita in our email exchanges, his update patches
will be sent to u-boot/kernel/OpenSBI later.

Update is based on this with changes:
https://github.com/riscv-software-src/opensbi/blob/master/docs/pmu_support.md

Signed-off-by: Xin Zhang <zhangxin.xa@gmail.com>